add demos jwt & add demo preview

+++ server/src/controllers/demo.js
@@ -3,7 +3,10 @@ import models from '../models'
 const Demo = models.Demo
 const demo = {}
 
+// TODO 给传参消毒
+
 demo.list = (req, res, next) => {
+  const userId = req.user.userId
   let countPerPage = 10, currentPage = req.query.page || 1
   if (!(/^[0-9]*$/).test(currentPage)) {
     currentPage = 1
@@ -11,7 +14,8 @@ demo.list = (req, res, next) => {
   currentPage = Number(currentPage)
   Demo.findAndCountAll({
     limit: countPerPage,
-    offset: countPerPage * (currentPage - 1)
+    offset: countPerPage * (currentPage - 1),
+    where: {UserId: userId}
   }).then((result) => {
     const data = {
       msg: '获取所有demo成功！',
@@ -29,8 +33,9 @@ demo.list = (req, res, next) => {
 }
 
 demo.get = (req, res, next) => {
+  const userId = req.user.userId
   const { id } = req.params
-  Demo.findOne({where: {id}}).then((demo) => {
+  Demo.findOne({where: {id,UserId: userId}}).then((demo) => {
     res.json({
       msg: '查询demo成功！',
       success: true,
@@ -42,8 +47,9 @@ demo.get = (req, res, next) => {
 }
 
 demo.create = (req, res, next) => {
-  let {img, name, material, year, color, introduce, voice, AuthorId} = req.body
-  Demo.create({img, name, material, year, color, introduce, voice, AuthorId}).then((demo) => {
+  const userId = req.user.userId
+  let {template, content} = req.body
+  Demo.create({template,content, UserId: userId}).then((demo) => {
     const data = {
       msg: '新建demo成功！',
       success: true,
@@ -55,9 +61,10 @@ demo.create = (req, res, next) => {
 }
 
 demo.update = (req, res) => {
+  const userId = req.user.userId
   const { id } = req.params
-  const { img, name, material, year, color, introduce, voice, AuthorId } = req.body
-  Demo.update({img, name, material, year, color, introduce, voice, AuthorId}, {where: {id}}).then(() => {
+  const { template, content } = req.body
+  Demo.update({template,content}, {where: {id,UserId: userId}}).then(() => {
     const data = {
       msg: '修改demo成功！',
       success: true
@@ -67,12 +74,22 @@ demo.update = (req, res) => {
 }
 
 demo.remove = (req, res, next) => {
+  const userId = req.user.userId
   const { id } = req.params
-  Demo.destroy({where: {id}}).then(() => {
+  Demo.destroy({where: {id,UserId: userId}}).then(() => {
     res.json({
       msg: '删除demo成功！',
       success: true
     })
+  }).catch((err) => {
+    next(err)
+  })
+}
+
+demo.preview = (req, res, next) => {
+  const { id } = req.params
+  Demo.findOne({where: {id}}).then((demo) => {
+    res.render(demo.template, JSON.parse(demo.content))
   }).catch((err) => {
     next(err)
   })
